Big Refactoring
-Add Folder structure
-Changed gitignore
-Build library and executable file
